pyproject.toml
[project]
name = "py-code-mode"
version = "0.1.0"
description = "Code mode for LLM agents - tools as Python SDK bindings via Jupyter execution"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "actae0n" }
]
keywords = ["llm", "agents", "tools", "jupyter", "code-execution"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
    "numpy>=1.26",
    "sentence-transformers>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]
jupyter = [
    "nbclient>=0.10",
    "nbformat>=5.10",
    "jupyter-client>=8.6",
]
chromadb = [
    "chromadb>=0.5",
]
redis = [
    "redis>=5.0",
]
mcp = [
    "mcp>=1.0",
]
http = [
    "aiohttp>=3.9",
]
container = [
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "httpx>=0.25",
    "docker>=7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/py_code_mode"]

[tool.hatch.build.targets.sdist]
include = ["src/py_code_mode"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
    "redis>=7.1.0",
    "mcp[cli]>=1.24.0",
    "aiohttp>=3.13.2",
    "fastapi>=0.124.4",
    "httpx>=0.28.1",
    "uvicorn>=0.38.0",
    "docker>=7.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
